// examin_rest_api_plus_frontend/examin-web-front-master/src/app/models/university.model.ts
export class University {
    universityId: number;
    course?: string;

    sscBoardName?: string;
    sscPercentage?: string;
    sscPassingYear?: string;
    sscCertificateNumber?: string;
    
    hscBoardName?: string;
    hscPercentage?: string;
    hscPassingYear?: string;
    hscCertificateNumber?: string;
    
    graduationBoardName?: string;
    graduationPercentage?: string;
    graduationPassingYear?: string;
    graduationCertificateNumber?: string;
    
    postGraduationBoardName?: string;
    postGraduationPercentage?: string;
    postGraduationPassingYear?: string;
    postGraduationCertificateNumber?: string;
}

// TODO: cleanup model if not in use